requirements.txt
pymupdf
sentence-transformers
faiss-cpu
groq
gradio
numpy